highway_env_GE_DQN_CLMHA/requirements.txt
torch
numpy
gym
highway-env
stable-baselines3
opencv-python
PyYAML
matplotlib
